+++ osmapi/note.py
@@ -91,7 +91,7 @@ class NoteMixin:
         Returns the updated note.
 
         If no authentication information are provided,
-        `OsmApi.UsernamePasswordMissingError` is raised.
+        `OsmApi.AuthenticationMissingError` is raised.
         """
         path = f"/api/0.6/notes/{note_id}/close"
         return self._note_action(path, comment, optional_auth=False)
@@ -105,7 +105,7 @@ class NoteMixin:
         Returns the updated note.
 
         If no authentication information are provided,
-        `OsmApi.UsernamePasswordMissingError` is raised.
+        `OsmApi.AuthenticationMissingError` is raised.
 
         If the requested element has been deleted,
         `OsmApi.ElementDeletedApiError` is raised.