+++ osmapi/relation.py
@@ -47,7 +47,7 @@ class RelationMixin:
         Creates a relation based on the supplied `relation_data` dict.
 
         If no authentication information are provided,
-        `OsmApi.UsernamePasswordMissingError` is raised.
+        `OsmApi.AuthenticationMissingError` is raised.
 
         If the supplied information contain an existing relation,
         `OsmApi.OsmTypeAlreadyExistsError` is raised.
@@ -67,7 +67,7 @@ class RelationMixin:
         Updates relation with the supplied `relation_data` dict.
 
         If no authentication information are provided,
-        `OsmApi.UsernamePasswordMissingError` is raised.
+        `OsmApi.AuthenticationMissingError` is raised.
 
         If there is no open changeset,
         `OsmApi.NoChangesetOpenError` is raised.
@@ -84,7 +84,7 @@ class RelationMixin:
         Delete relation with `relation_data`.
 
         If no authentication information are provided,
-        `OsmApi.UsernamePasswordMissingError` is raised.
+        `OsmApi.AuthenticationMissingError` is raised.
 
         If there is no open changeset,
         `OsmApi.NoChangesetOpenError` is raised.